docs: sd-select docs [skip chromatic]  (#1289)
