+++ packages/components/src/components/select/select.stories.ts
@@ -2,40 +2,30 @@
 import '../../solid-components';
 import { html } from 'lit-html';
 import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../../scripts/storybook/helper';
-import { userEvent } from '@storybook/test';
-import { waitUntil } from '@open-wc/testing-helpers';
 import type { ConstantDefinition } from '../../../scripts/storybook/helper';
 
 const { argTypes, parameters } = storybookDefaults('sd-select');
 const { generateTemplate } = storybookTemplate('sd-select');
 const { overrideArgs } = storybookHelpers('sd-select');
 
-// Reusable Constants
-const twoOptionsConstant: ConstantDefinition = {
-  type: 'slot',
-  name: 'default',
-  value: '<sd-option value="option-1">Option 1</sd-option><sd-option value="option-2">Option 2</sd-option>'
-};
 const threeOptionsConstant: ConstantDefinition = {
   type: 'slot',
   name: 'default',
   value:
     '<sd-option value="option-1">Option 1</sd-option><sd-option value="option-2">Option 2</sd-option><sd-option value="option-3">Option 3</sd-option>'
 };
-const fiveOptionsConstant: ConstantDefinition = {
-  type: 'slot',
-  name: 'default',
-  value:
-    '<sd-option value="option-1">Option 1</sd-option><sd-option value="option-2">Option 2</sd-option><sd-option value="option-3">Option 3</sd-option><sd-option value="option-4">Option 4</sd-option><sd-option value="option-5">Option 5</sd-option>'
-};
-const clearableConstant: ConstantDefinition = { type: 'attribute', name: 'clearable', value: true };
-const multipleConstant: ConstantDefinition = { type: 'attribute', name: 'multiple', value: true };
-const helpTextConstant: ConstantDefinition = { type: 'attribute', name: 'help-text', value: 'help-text' };
 const labelConstant: ConstantDefinition = { type: 'attribute', name: 'label', value: 'Label' };
 
-// Stories
+/**
+ * Used to choose items from a menu of predefined options.
+ *
+ * **Related templates:**
+ * - [Select Grouping](?path=/docs/templates-select--docs)
+ * - [Select with Tooltip](?path=/docs/templates-select--docs#select%20with%20tooltip)
+ */
 export default {
   title: 'Components/sd-select',
+  tags: ['!dev'],
   component: 'sd-select',
   args: overrideArgs([
     threeOptionsConstant,
@@ -71,600 +61,331 @@ export default {
 
 export const Default = {
   render: (args: any) => {
-    return html`<div class="h-[260px] w-[420px]">${generateTemplate({ args })}</div>`;
+    return html`<div class="h-[260px] w-[400px]">${generateTemplate({ args })}</div>`;
   }
 };
 
 /**
- * Use the `size` attribute to change a select’s size. It will cascade to slotted `sd-option` elements.
+ * Use the `size` attribute the size. It will cascade to slotted `sd-option` elements.
  */
 
-export const SizeMultiple = {
-  name: 'Size x Multiple',
-  parameters: {
-    controls: {
-      exclude: ['open-attr']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
+export const Size = {
+  render: () => html`
+    <div class="flex gap-12 h-[500px]">
+      <sd-select size="lg" label="Label" placeholder="Large" placement="bottom" value="" class="test">
+        <sd-option value="option-1" size="lg">Option 1</sd-option>
+        <sd-option value="option-2" size="lg">Option 2</sd-option>
+        <sd-option value="option-3" size="lg">Option 3</sd-option>
+        <sd-option value="option-4" size="lg">Option 4</sd-option>
+        <sd-option value="option-5" size="lg">Option 5</sd-option>
+      </sd-select>
 
-    return html`<div class="h-[340px]">
-      ${generateTemplate({
-        options: {
-          classes: 'w-full'
-        },
-        axis: {
-          x: {
-            type: 'attribute',
-            name: 'size'
-          },
-          y: {
-            type: 'attribute',
-            name: 'useTags',
-            values: [false, true]
-          }
-        },
-        constants: [
-          fiveOptionsConstant,
-          multipleConstant,
-          { type: 'attribute', name: 'value', value: 'option-1 option-2 option-3 option-4' }
-        ],
-        args
-      })}
-    </div>`;
-  }
+      <sd-select size="md" label="Label" placeholder="Medium" placement="bottom" value="" class="test">
+        <sd-option value="option-1" size="md">Option 1</sd-option>
+        <sd-option value="option-2" size="md">Option 2</sd-option>
+        <sd-option value="option-3" size="md">Option 3</sd-option>
+        <sd-option value="option-4" size="md">Option 4</sd-option>
+        <sd-option value="option-5" size="md">Option 5</sd-option>
+      </sd-select>
+
+      <sd-select size="sm" label="Label" placeholder="Small" placement="bottom" value="">
+        <sd-option value="option-1" size="sm">Option 1</sd-option>
+        <sd-option value="option-2" size="sm">Option 2</sd-option>
+        <sd-option value="option-3" size="sm">Option 3</sd-option>
+        <sd-option value="option-4" size="sm">Option 4</sd-option>
+        <sd-option value="option-5" size="sm">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
 };
 
 /**
- * To allow multiple options to be selected, use the `multiple` attribute. It’s a good practice to use `clearable` when this option is enabled. To use the checkbox with tags variant, set the `useTags` variant to `true`.  To set multiple values at once, set value to a space-delimited list of values.  The preferred placement of the select’s listbox can be set with the `placement` attribute. Note that the actual position may vary to ensure the panel remains in the viewport. Valid placements are `top` and `bottom`.
+ * Use the `placement` attribute to define where the select panel should appear.
  */
 
-export const DisabledMultiple = {
-  name: 'Disabled x Multiple',
-  parameters: {
-    controls: {
-      exclude: ['open']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
+export const Placement = {
+  render: () => html`
+    <div class="flex items-center gap-12 h-[500px]">
+      <sd-select class="self-baseline" size="lg" label="Label" placement="bottom" value="">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
 
-    return html`<div class="h-[340px] w-full">
-      ${generateTemplate({
-        options: {
-          classes: 'w-full [&>tbody>tr>td]:w-[50%]'
-        },
-        axis: {
-          y: {
-            type: 'attribute',
-            name: 'useTags',
-            values: [false, true]
-          },
-          x: {
-            type: 'attribute',
-            name: 'disabled',
-            values: [false, true]
-          }
-        },
-        constants: [
-          clearableConstant,
-          multipleConstant,
-          fiveOptionsConstant,
-          {
-            type: 'attribute',
-            name: 'value',
-            value: 'option-1 option-2 option-3 option-4'
-          }
-        ],
-        args
-      })}
-    </div>`;
-  }
+      <sd-select size="lg" label="Label" placement="top" value="">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
 };
 
 /**
- * `sd-select` with valid and invalid styles.
+ * Use the `label` attribute to give the select element an accessible label.<br />For labels that contain HTML, use the `label` slot instead.
  */
 
-export const ValidInvalid = {
-  name: 'Valid x Invalid',
-  parameters: {
-    controls: {
-      exclude: ['label', 'open-attr', 'required', 'default', 'useTags', 'multiple', 'max-options-visible']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
+export const Label = {
+  render: () => html`
+    <div class="flex gap-12 h-[500px]">
+      <sd-select size="lg" label="Label Attribute" placement="bottom" value="">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
 
-    return html`<form class="h-[260px] w-full flex gap-4">
-      ${generateTemplate({
-        options: {
-          classes: 'w-full [&>tbody>tr>td]:align-top'
-        },
-        axis: {
-          y: {
-            type: 'attribute',
-            name: 'useTags',
-            values: [false, true]
-          },
-          x: {
-            type: 'attribute',
-            name: 'value',
-            values: ['option-1 option-2', '']
-          }
-        },
-        constants: [
-          twoOptionsConstant,
-          labelConstant,
-          multipleConstant,
-          { type: 'attribute', name: 'required', value: true }
-        ],
-        args
-      })}
-      <sd-button class="hidden" type="submit">Submit</sd-button>
-    </form>`;
-  },
-  play: async ({ canvasElement }: { canvasElement: HTMLUnknownElement }) => {
-    const el = canvasElement.querySelector('sd-button');
-    await waitUntil(() => el?.shadowRoot?.querySelector('button'));
-    await userEvent.type(el!.shadowRoot!.querySelector('button')!, '{return}', { pointerEventsCheck: 0 });
-  }
+      <sd-select size="lg" placement="bottom" value="">
+        <div slot="label" class="text-lg">Label Slot</div>
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
 };
 
 /**
- * Shows available slots. The `label` and `help-text` slots will overwrite their corresponding attributes.
+ * Use the `placeholder` attribute to show a placeholder when no option is selected.
  */
 
-export const Slots = {
-  parameters: {
-    controls: {
-      exclude: ['open-attr']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
-
-    return html`
-      ${['default', 'label', 'clear-icon', 'expand-icon', 'help-text'].map(slot =>
-        generateTemplate({
-          axis: {
-            x: {
-              type: 'slot',
-              name: slot,
-              title: 'slot=...',
-              values: [
-                {
-                  value:
-                    slot === 'default'
-                      ? `<div class="slot slot--border slot--background h-8 w-full"></div>`
-                      : `<div slot='${slot}' class="slot slot--border slot--background h-6 ${
-                          slot === 'label' || slot === 'help-text' ? 'w-20' : 'w-6'
-                        }"></div>`,
-                  title: slot
-                }
-              ]
-            }
-          },
-          constants: [
-            { type: 'template', name: 'width', value: '<div style="width: 300px">%TEMPLATE%</div>' },
-            clearableConstant,
-            labelConstant,
-            helpTextConstant,
-            twoOptionsConstant
-          ],
-          args
-        })
-      )}
-    `;
-  }
+export const Placeholder = {
+  render: () => html`
+    <div class="w-[400px] h-[500px]">
+      <sd-select size="lg" label="Label" placeholder="Placeholder example" placement="bottom" value="">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
 };
 
 /**
- * Use the `form-control`,
-  `form-control-label`,
-  `form-control-input`,
-  `form-control-help-text`,
-  `combobox`,
-  `display-input`,
-  `listbox`,
-  `tags`,
-  `tag`,
-  `tag__base`,
-  `tag__content`,
-  `tag__removable-indicator`,
-  `clear-button`, and
-  `expand-icon` part selectors to customize the select component.
+ * Use the `disabled` attribute to disable the select.
  */
 
-const partsArr = [
-  'form-control',
-  'form-control-label',
-  'form-control-input',
-  'form-control-help-text',
-  'combobox',
-  'display-input',
-  'listbox',
-  'tags',
-  'tag',
-  'tag__base',
-  'tag__content',
-  'tag__removable-indicator',
-  'clear-button',
-  'expand-icon'
-];
-
-export const Parts = {
-  parameters: {
-    controls: {
-      exclude: [
-        'open-attr',
-        'form-control',
-        'form-control-label',
-        'form-control-input',
-        'form-control-help-text',
-        'combobox',
-        'display-input',
-        'listbox',
-        'tags',
-        'tag',
-        'tag__base',
-        'tag__content',
-        'tag__removable-indicator',
-        'clear-button',
-        'expand-icon'
-      ]
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
-
-    return generateTemplate({
-      axis: {
-        x: { type: 'attribute', name: 'useTags' },
-        y: {
-          type: 'template',
-          name: 'sd-select::part(...){outline: solid 2px red}',
-          values: partsArr.map(part => {
-            return {
-              title: part,
-              value: `<style>#part-${part} sd-select::part(${part}){outline: solid 2px red; outline-offset: 2px} ::part(popup__content){overflow-y: visible} ::part(tag__removable-indicator){display: block} ::part(tag__content){display: block}</style><div id="part-${part}">%TEMPLATE%</div>`
-            };
-          })
-        }
-      },
-      constants: [
-        { type: 'attribute', name: 'useTags', value: true },
-        { type: 'attribute', name: 'value', value: 'option-1 option-2' },
-        clearableConstant,
-        helpTextConstant,
-        labelConstant,
-        multipleConstant
-      ],
-      args
-    });
-  }
+export const Disabled = {
+  render: () => html`
+    <div class="w-[400px]">
+      <sd-select size="lg" label="Label" placeholder="Disabled" placement="bottom" value="" disabled>
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
 };
 
 /**
- * Per default the select will indicate an error state when the input is invalid. Use the `style-on-valid` attribute to indicate a valid state as well.
+ * Use the `help-text` attribute to provide additional context or instructions.
  */
 
-export const StyleOnValid = {
-  parameters: {
-    controls: {
-      exclude: ['open-attr']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
+export const HelpText = {
+  render: () => html`
+    <div class="flex gap-12 h-[500px]">
+      <sd-select size="lg" label="Label" placement="bottom" value="" help-text="Help text Attribute">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
 
-    return html`<div class="h-[340px]">
-      ${generateTemplate({
-        options: {
-          classes: 'w-full'
-        },
-        axis: {
-          x: {
-            type: 'attribute',
-            name: 'style-on-valid'
-          }
-        },
-        constants: [fiveOptionsConstant, { type: 'attribute', name: 'value', value: '' }],
-        args
-      })}
-    </div>`;
-  },
+      <sd-select size="lg" label="Label" placement="bottom" value="">
+        <div slot="help-text" class="text-lg">Help text Slot</div>
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
+};
 
-  play: async ({ canvasElement }: { canvasElement: HTMLUnknownElement }) => {
-    await Promise.all([customElements.whenDefined('sd-select'), customElements.whenDefined('sd-option')]).then(
-      async () => {
-        const els = canvasElement.querySelectorAll('sd-select');
+/**
+ * Use the `clearable` attribute to add a clear button that removes the selected value.
+ */
 
-        for (const el of els) {
-          await waitUntil(() => el?.shadowRoot?.querySelector('input'));
-          await userEvent.click(el.shadowRoot!.querySelector('input')!);
-          await userEvent.click(el.querySelector('sd-option')!);
-        }
+export const Clearable = {
+  render: () => html`
+    <div class="w-[400px] h-[500px]">
+      <sd-select size="lg" label="Label" placement="bottom" clearable="" value="option-1">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
+};
 
-        // tab to next element to loose focus
-        await userEvent.tab();
+/**
+ * Use the `multiple` attribute to allow multiple options to be selected.
+ */
+
+export const Multiple = {
+  render: () => html`
+    <div class="w-[400px] h-[500px]">
+      <sd-select size="lg" label="Label" placement="bottom" multiple="" value="option-1 option-2">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
+};
+
+/**
+ * Use the `useTags` attribute to display selected options as tags using the `sd-tag` component.
+ *
+ * **Hint:** it requires the `multiple` attribute to be set.
+ */
+
+export const useTags = {
+  render: () => html`
+    <div class="w-[400px] h-[500px]">
+      <sd-select size="lg" label="Label" placement="bottom" multiple="" value="option-1 option-2" useTags>
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
+};
+
+/**
+ * Use the `max-options-visible` attribute to define the maximum number of selected options that will be visible.
+ *
+ * **Hint:** it requires the `multiple` and `useTags` attributes to be set.<br />After the maximum number of options is reached, the select will display a message indicating how many more options are selected.<br />To remove the limit, set the attribute to `0`.
+ */
+
+export const MaxOptionsVisible = {
+  render: () => html`
+    <div class="w-[400px] h-[500px]">
+      <sd-select
+        size="lg"
+        label="Label"
+        placement="bottom"
+        multiple=""
+        value="option-1 option-2 option-3"
+        useTags
+        max-options-visible="2"
+      >
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
+};
+
+/**
+ * Use the `required` attribute to mark the element as required. This can be used for form validation purposes.
+ */
+
+export const Required = {
+  render: () => html`
+    <div class="w-[400px] h-[400px]">
+      <sd-select size="lg" label="Required" placeholder="Please Select" placement="bottom" value="" required="">
+        <sd-option value="option-1">Option 1</sd-option>
+        <sd-option value="option-2">Option 2</sd-option>
+        <sd-option value="option-3">Option 3</sd-option>
+        <sd-option value="option-4">Option 4</sd-option>
+        <sd-option value="option-5">Option 5</sd-option>
+      </sd-select>
+    </div>
+  `
+};
+
+/**
+ * The component gets `valid` state when the input is valid.<br />Use the `style-on-valid` attribute to automatically indicate and show a valid state.
+ */
+
+export const Valid = {
+  render: () => html`
+    <style>
+      sd-select.valid-example::part(display-input) {
+        color: #181818;
       }
-    );
-  }
+    </style>
+    <div class="w-[400px] h-[400px]">
+      <sd-select
+        size="lg"
+        placement="bottom"
+        label="Label"
+        placeholder="Please Select"
+        style-on-valid=""
+        value="option-1"
+        class="valid-example"
+      >
+        <sd-option class="option" value="option-1">Option 1</sd-option>
+        <sd-option class="option" value="option-2">Option 2</sd-option>
+        <sd-option class="option" value="option-3">Option 12</sd-option>
+      </sd-select>
+    </div>
+
+    <script>
+      var validSelect = document.querySelector('.valid-example');
+      var options = validSelect.querySelectorAll('.option');
+      setTimeout(() => {
+        validSelect.checkValidity();
+        validSelect.reportValidity();
+      }, 500);
+    </script>
+  `
 };
 
 /**
- * `sd-select` is fully accessibile via keyboard.
+ * The component gets `invalid` state when the form is not valid.
  */
 
-export const Mouseless = {
-  parameters: {
-    controls: {
-      exclude: ['open-attr', 'value', 'default']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
+export const Invalid = {
+  render: () => html`
+    <div class="w-[400px] h-[300px]">
+      <sd-select
+        size="lg"
+        placement="bottom"
+        label="Label"
+        placeholder="Please Select"
+        style-on-valid=""
+        value=""
+        required=""
+        clearable=""
+        class="invalid-example"
+      >
+        <sd-option class="option" value="option-1">Option 1</sd-option>
+        <sd-option class="option" value="option-2">Option 2</sd-option>
+        <sd-option class="option" value="option-3">Option 3</sd-option>
+      </sd-select>
+    </div>
 
-    return html`<div class="mouseless h-[260px] w-full flex gap-4">
-      ${generateTemplate({
-        constants: [twoOptionsConstant, { type: 'attribute', name: 'label', value: 'Default' }],
-        args
-      })}
-      ${generateTemplate({
-        constants: [twoOptionsConstant, multipleConstant, { type: 'attribute', name: 'label', value: 'Multiple' }],
-        args
-      })}
-      ${generateTemplate({
-        constants: [
-          twoOptionsConstant,
-          multipleConstant,
-          { type: 'attribute', name: 'useTags', value: true },
-          { type: 'attribute', name: 'label', value: 'Multiple w/ Tags' }
-        ],
-        args
-      })}
-    </div>`;
-  },
-
-  play: async ({ canvasElement }: { canvasElement: HTMLUnknownElement }) => {
-    const el = canvasElement.querySelector('.mouseless sd-select');
-    await waitUntil(() => el?.shadowRoot?.querySelector('input'));
-    el?.shadowRoot?.querySelector('input')!.focus();
-  }
-};
-
-/**
- * Use `<sl-divider>` to group listbox items visually. You can also use `<small>` to provide labels, but they won’t be announced by most assistive devices.
- */
-
-export const SampleGroupingOptions = {
-  name: 'Sample: Grouping Options',
-  parameters: {
-    controls: {
-      exclude: ['open-attr', 'default']
-    }
-  },
-  render: (args: any) => {
-    return html`<div class="h-[290px] w-[420px]">
-      ${generateTemplate({
-        constants: [
-          {
-            type: 'slot',
-            name: 'default',
-            value:
-              '<div class="text-black px-4 font-bold">Nisi eu excepteur anim esse</div><sd-option value="option-1">Option 1</sd-option><sd-option value="option-2">Option 2</sd-option><sd-divider class="mb-2"></sd-divider><div class="text-black px-4 font-bold">Nisi eu excepteur anim esse</div><sd-option value="option-3">Option 3</sd-option>'
-          }
-        ],
-        args
-      })}
-    </div>`;
-  }
-};
-
-/**
- * Demonstrates the form behavior with validation styles when the `required` attribute is set to `true`.
- */
-
-export const SampleForm = {
-  name: 'Sample: Form',
-  parameters: {
-    controls: {
-      exclude: ['open-attr', 'label', 'name', 'useTags', 'value', 'multiple', 'max-options-visible', 'default']
-    }
-  },
-  render: (args: { 'open-attr'?: string }) => {
-    delete args['open-attr'];
-
-    const sharedConstants: ConstantDefinition[] = [
-      { type: 'attribute', name: 'form', value: 'testForm' },
-      { type: 'attribute', name: 'clearable', value: true },
-      { type: 'attribute', name: 'required', value: true },
-      twoOptionsConstant
-    ];
-
-    return html`
-      <form action="" method="get" id="testForm" name="testForm" class="w-[370px]">
-        <div class="mb-6">
-          ${generateTemplate({
-            constants: [
-              ...sharedConstants,
-              { type: 'attribute', name: 'label', value: 'Required' },
-              { type: 'attribute', name: 'name', value: 'required field' }
-            ],
-            args
-          })}
-        </div>
-        <div class="mb-6">
-          ${generateTemplate({
-            constants: [
-              ...sharedConstants,
-              { type: 'attribute', name: 'label', value: 'Required Multiple' },
-              { type: 'attribute', name: 'name', value: 'required multiple field' },
-              multipleConstant
-            ],
-            args
-          })}
-        </div>
-        <div class="mb-8">
-          ${generateTemplate({
-            constants: [
-              ...sharedConstants,
-              { type: 'attribute', name: 'label', value: 'Required Multiple w/ Tags' },
-              { type: 'attribute', name: 'name', value: 'required multiple tags field' },
-              multipleConstant,
-              { type: 'attribute', name: 'useTags', value: true }
-            ],
-            args
-          })}
-        </div>
-        <sd-button type="submit">Submit</sd-button>
-      </form>
-      <script>
-        function handleSubmit(event) {
-          const form = document.querySelector('#testForm');
-          const sdSelects = Array.from(document.querySelectorAll('sd-select'));
-
-          const isValid = sdSelect => sdSelect.checkValidity();
-
-          if (sdSelects.every(isValid)) {
-            event.preventDefault(); // Prevent the default form submission behavior
-
-            const formData = new FormData(form);
-            const formValues = Object.fromEntries(formData);
-
-            alert('Form submitted successfully with the following values: ' + JSON.stringify(formValues, null, 2));
-          }
-        }
-
-        document.querySelector('#testForm').addEventListener('submit', handleSubmit);
-      </script>
-    `;
-  }
-};
-
-/**
- * 1. You can use the `setCustomValidity` method to set a custom validation message. This will override any native validation messages.
- * 2. Set an empty string to clear the custom validity and make the input valid.
- * 3. To show the validation message, call the `reportValidity` method. Originally this would show a native validation bubble, but we show the error messages inline.
- */
-
-export const setCustomValidity = {
-  parameters: {
-    chromatic: { disableSnapshot: true }
-  },
-  render: () => {
-    return html`
-      <!-- block submit and show alert instead -->
-      <form id="validationForm" class="flex flex-col gap-2">
-        <sd-select id="custom-input" style-on-valid>
-          <sd-option value="option-1">Option 1</sd-option>
-          <sd-option value="option-2">Option 2</sd-option>
-          <sd-option value="option-3">Option 3</sd-option>
-        </sd-select>
-        <div>
-          <sd-button type="submit">Submit</sd-button>
-          <sd-button id="error-button" variant="secondary">Set custom error</sd-button>
-          <sd-button id="success-button" variant="secondary">Set success</sd-button>
-          <sd-button type="reset" variant="secondary">Reset</sd-button>
-        </div>
-      </form>
-      <script type="module">
-        // Wait for custom elements to be defined
-        await Promise.all([
-          customElements.whenDefined('sd-select'),
-          customElements.whenDefined('sd-button'),
-          customElements.whenDefined('sd-option')
-        ]).then(() => {
-          const form = document.getElementById('validationForm');
-          const input = document.getElementById('custom-input');
-          const setErrorButton = document.getElementById('error-button');
-          const setSuccessButton = document.getElementById('success-button');
-
-          // Initial error
-          const errorMessage = \`This is an initial custom error (\${new Date().toLocaleTimeString()})\`;
-          input.setCustomValidity(errorMessage);
-          input.reportValidity();
-
-          // Show error message
-          setErrorButton.addEventListener('click', () => {
-            const errorMessage = \`This is a new custom error (\${new Date().toLocaleTimeString()})\`;
-            input.setCustomValidity(errorMessage);
-            input.reportValidity();
-          });
-
-          // Show success message
-          setSuccessButton.addEventListener('click', () => {
-            input.setCustomValidity(''); // Clear custom validity
-            input.reportValidity();
-          });
-
-          form.addEventListener('submit', event => {
-            event.preventDefault();
-            alert('All fields are valid!');
-          });
-        });
-      </script>
-    `;
-  }
-};
-
-/**
- * Demonstrates a form containing all existing Solid form elements.
- */
-
-export const SolidForm = {
-  name: 'Sample: Solid Form',
-  parameters: {
-    controls: {
-      include: []
-    }
-  },
-  render: () => {
-    return html`
-      <form action="" method="get" id="testForm" name="testForm" class="">
-        <h1 class="text-lg text-white bg-primary p-4">Solid Form</h1>
-        <div class="[&>:nth-child(even)]:bg-neutral-100 [&>*]:p-4">
-          <sd-checkbox form="testForm" name="field 1" required>Field 1</sd-checkbox>
-          <sd-input form="testForm" name="field 2" label="Field 2" required></sd-input>
-          <sd-select form="testForm" name="field 3" label="Field 3" required
-            ><sd-option value="option-1">Option 1</sd-option><sd-option value="option-2">Option 2</sd-option>
-            <sd-option value="option-3">Option 3</sd-option><sd-option value="option-4">Option 4</sd-option
-            ><sd-option value="option-5">Option 5</sd-option><sd-option value="option-6">Option 6</sd-option
-            ><sd-option value="option-7">Option 7</sd-option></sd-select
-          >
-          <sd-radio-group form="testForm" name="field 4" label="Field 4" required
-            ><sd-radio value="option-1">Option 1</sd-radio><sd-radio value="option-2">Option 2</sd-radio>
-            <sd-radio value="option-3">Option 3</sd-radio></sd-radio-group
-          >
-          <sd-radio-group value="option-1" form="testForm" name="field 5" label="Field 5" required
-            ><sd-radio-button value="option-1">Option 1</sd-radio-button
-            ><sd-radio-button value="option-2">Option 2</sd-radio-button>
-            <sd-radio-button value="option-3">Option 3</sd-radio-button></sd-radio-group
-          >
-          <sd-switch form="testForm" name="field 6" required>Field 6</sd-switch>
-          <sd-textarea form="testForm" name="field 7" label="Field 7" required></sd-textarea>
-        </div>
-        <sd-button class="my-4" type="submit">Submit</sd-button>
-        <sd-button class="my-4" type="reset">Reset</sd-button>
-      </form>
-
-      <script>
-        function handleSubmit(event) {
-          const form = document.querySelector('#testForm');
-
-          const formData = new FormData(form);
-          const formValues = Object.fromEntries(formData);
-
-          if (form.reportValidity()) {
-            event.preventDefault(); // Prevent the default form submission behavior
-            alert('Form submitted with the following values: ' + JSON.stringify(formValues, null, 2));
-          }
-        }
-
-        document.querySelector('#testForm').addEventListener('submit', handleSubmit);
-      </script>
-    `;
-  }
+    <script>
+      var invalidSelect = document.querySelector('.invalid-example');
+      setTimeout(() => {
+        invalidSelect.checkValidity();
+        invalidSelect.reportValidity();
+        invalidSelect.setCustomValidity('Error text');
+      }, 500);
+    </script>
+  `
 };
