+++ packages/components/src/components/select/select.ts
@@ -121,20 +121,23 @@ export default class SdSelect extends SolidElement implements SolidFormControl {
   /** The select's size. */
   @property({ reflect: true }) size: 'lg' | 'md' | 'sm' = 'lg';
 
+  /**
+   * The preferred placement of the select's menu. Note that the actual placement may vary as needed to keep the listbox
+   * inside of the viewport.
+   */
+  @property({ reflect: true }) placement: 'top' | 'bottom' = 'bottom';
+
   /** The select's label. If you need to display HTML, use the `label` slot instead. */
   @property() label = '';
 
   /** Placeholder text to show as a hint when the select is empty. */
   @property() placeholder = this.localize.term('selectDefaultPlaceholder');
 
+  /** Disables the select control. */
+  @property({ type: Boolean, reflect: true }) disabled = false;
+
   /** The select's help text. If you need to display HTML, use the `help-text` slot instead. */
   @property({ attribute: 'help-text' }) helpText = '';
-
-  /**
-   * The preferred placement of the select's menu. Note that the actual placement may vary as needed to keep the listbox
-   * inside of the viewport.
-   */
-  @property({ reflect: true }) placement: 'top' | 'bottom' = 'bottom';
 
   /**
    * The actual current placement of the select's menu sourced from `sd-popup`.
@@ -144,9 +147,6 @@ export default class SdSelect extends SolidElement implements SolidFormControl {
 
   /** Adds a clear button when the select is not empty. */
   @property({ type: Boolean }) clearable = false;
-
-  /** Disables the select control. */
-  @property({ type: Boolean, reflect: true }) disabled = false;
 
   /** Allows more than one option to be selected. */
   @property({ type: Boolean, reflect: true }) multiple = false;
